requirements.txt
crewai>=0.1.0
pandas>=1.5.0
yfinance>=0.2.0
scikit-learn>=1.2.0
joblib>=1.2.0
alpaca-trade-api>=2.0.0
python-dotenv>=0.21.0